# backend/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "properties": {
                  "key": {
                    "type": "string"
                  },
                  "params": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  }
                },
                "required": [
                  "key"
                ],
                "type": "object"
              },
              "type": "array"
            },
            "type": "object"
          },
          "instance": {
            "type": "string"
          },
          "messageKey": {
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "code",
          "messageKey",
          "correlationId"
        ],
        "type": "object"
      },
...
      "PublicSite": {
        "properties": {
          "defaultLocale": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "enabledLocales": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "showPlatformAttribution": {
            "type": "boolean"
          },
          "theme": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PublicTheme"
              }
            ]
          }
        },
        "required": [
          "displayName",
          "enabledLocales",
          "defaultLocale",
          "theme",
          "showPlatformAttribution"
        ],
        "type": "object"
      },
...
      "PublicTheme": {
        "properties": {
          "background": {
            "type": "string"
          },
          "foreground": {
            "type": "string"
          },
          "muted": {
            "type": "string"
          },
          "primary": {
            "type": "string"
          }
        },
        "required": [
          "primary",
          "background",
          "foreground",
          "muted"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LearnStack API",
    "version": "v1"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/public/site": {
      "get": {
        "description": "Anonymous host-scoped read through the read-only pipeline. Unknown query keys and repeated single-valued keys are refused. No response caching or validators.",
        "operationId": "GetPublicSite",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSite"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "validation_failed; the site bootstrap accepts no query parameters.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "not_found; unavailable host or hidden/untranslated/ineligible resource, without disclosure.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "method_not_allowed from routing.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "rate_limited from the anonymous limiter.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "internal_error from unexpected failures; no exception or query details.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "dependency_unavailable from invalid stored public configuration.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "PublicSite"
        ],
        "x-version-introduced": "v1"
      },
      "head": {
        "description": "Anonymous host-scoped read through the read-only pipeline. Unknown query keys and repeated single-valued keys are refused. No response caching or validators.",
        "operationId": "HeadPublicSite",
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "validation_failed; the site bootstrap accepts no query parameters.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "not_found; unavailable host or hidden/untranslated/ineligible resource, without disclosure.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "description": "method_not_allowed from routing.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited from the anonymous limiter.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "internal_error from unexpected failures; no exception or query details.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "dependency_unavailable from invalid stored public configuration.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "PublicSite"
        ],
        "x-version-introduced": "v1"
      }
    }
  },
...
}